+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 Released 2017-03-13
   - Merge: Make sure the output 'cartodb_id' is unique
   - Merge: Allow chained merge analyses
+  - TradeArea: Allow chained analyses
 
 ## 0.61.3
 Released 2017-03-12