+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 ## 0.61.4
-Released 2017-03-13
+Released 2017-03-14
   - Merge: Make sure the output 'cartodb_id' is unique
   - Merge: Allow chained merge analyses
   - TradeArea: Allow chained analyses
+  - Update request to 2.85.0
 
 ## 0.61.3
 Released 2017-03-12